Require a proper 'scope'
RFC 6749 Section 3.3 requires that the server either apply and document a default scope, or
fail the request. For this generic implementation the default is to fail when the model
does not provide a scope -- and leaves it to the model to decide whether and how it handles
a missing scope from the request.

This then also allows us to formally require the `scope` property on authorization
codes, refresh tokens, and access tokens, which the test code already implied.