WIP: Make the library pass a basic `install&&lint&&build` run
A lot of the changes here should be split into smaller pieces, and then get pulled into nkitku's v5.0-dev branch.

For possing the tests usually 'any' things got changed to their proper types, and the fallout was fixed. There is
a lot of these places left for future work, this commit only touched code that was leading to a failing test, usually
because of the changes to
* require the scope
* include the client/user in the generated tokens